careerpilot-ai/requirements.txt
streamlit
pdfplumber
python-dotenv
google-generativeai
